[tests] created test cases
[base_functionalities] handle the create/stup/teardown + other common functionalities which are used in all subtest classes

[test_endpoints_protection] created tests for login required and permissions checks based on different user roles.

[test_register_endpoint] test suite for verifying the register schema requirements and cases

[factories] cretead factories for User/Item/Order

[test_login_endpoint] verifying login schema requirements and email validator

[config] added config for testing env
[authenticator] common msg for invalid token
[app] refactor for testing env

[items] on delete method raises badrequest if None

[base_item_schema] removed last_update_by to be visibile from all also added
[base_item_crud] added test and validation for item CRUDs. TODO MOCK addSpeck

[schemas/response/item] manager schema has price + last_update_by filed

+++ resources/items.py
@@ -19,6 +19,7 @@ from schemas.response.item import (
 )
 from services.icecat.icecat_extractor import IceCatExtractor
 from utils.decorators import permission_required, validate_schema
+from werkzeug.exceptions import BadRequest
 
 
 class CreateItem(Resource):
@@ -36,6 +37,8 @@ class DeleteItem(Resource):
     @auth.login_required
     @permission_required(UserRole.manager)
     def delete(self, item_id: int):
+        if not item_id:
+            raise BadRequest(f"Item id cannot be Null or Zero!")
         ItemManager.delete_item(item_id)
         return {}, 204
